Add Float Support

+++ test/test_bytecode.cpp
@@ -20,7 +20,9 @@ std::vector<PythonOpcode> extractInstructionOpcodes(const std::vector<ByteCodeIn
     return opcodes;
 }
 
-TEST_CASE("Test Empty Bytecode") {
+TEST_CASE(
+        "Test Empty Bytecode"
+        ) {
     const std::string source; // Blank source
     const std::vector expectedOpcodes = {PythonOpcode::RESUME, PythonOpcode::LOAD_CONST, PythonOpcode::RETURN_VALUE};
     const ByteCodeModule bytecodeModule = compilePython(source, "<embedded>");
@@ -30,7 +32,9 @@ TEST_CASE("Test Empty Bytecode") {
 }
 
 
-TEST_CASE("Test Integer Assign Bytecode") {
+TEST_CASE(
+        "Test Integer Assign Bytecode"
+        ) {
     SECTION("Test Small Integer") {
         const std::string source = "x = 42";
         const std::vector expectedOpcodes = {PythonOpcode::RESUME, PythonOpcode::LOAD_SMALL_INT,
@@ -41,7 +45,7 @@ TEST_CASE("Test Integer Assign Bytecode") {
 
         REQUIRE(opcodes == expectedOpcodes);
 
-        const int* instructionInt = std::get_if<int>(&(bytecodeModule.instructions[1].argval));
+        const int64_t* instructionInt = std::get_if<int64_t>(&(bytecodeModule.instructions[1].argval));
         REQUIRE(instructionInt != nullptr);
         REQUIRE(*instructionInt == 42);
 
@@ -63,7 +67,9 @@ TEST_CASE("Test Integer Assign Bytecode") {
 }
 
 
-TEST_CASE("Test Function Bytecode") {
+TEST_CASE(
+        "Test Function Bytecode"
+        ) {
     const std::string source = "def func(x): return x + 1\nfunc(2)";
     const std::vector expectedOpcodes = {PythonOpcode::RESUME, PythonOpcode::LOAD_CONST, PythonOpcode::MAKE_FUNCTION,
                                          PythonOpcode::STORE_NAME, PythonOpcode::LOAD_NAME, PythonOpcode::PUSH_NULL,
@@ -75,13 +81,15 @@ TEST_CASE("Test Function Bytecode") {
     REQUIRE(opcodes == expectedOpcodes);
     REQUIRE(bytecodeModule.instructions[3].argrepr == "func");
 
-    const int* instructionInt = std::get_if<int>(&(bytecodeModule.instructions[6].argval));
+    const int64_t* instructionInt = std::get_if<int64_t>(&(bytecodeModule.instructions[6].argval));
     REQUIRE(instructionInt != nullptr);
     REQUIRE(*instructionInt == 2);
 }
 
 
-TEST_CASE("Test Builtin Bytecode") {
+TEST_CASE(
+        "Test Builtin Bytecode"
+        ) {
     SECTION("Test Print Builtin") {
         const std::string source = "print('hello')";
         const std::vector expectedOpcodes = {PythonOpcode::RESUME, PythonOpcode::LOAD_NAME,
@@ -111,7 +119,9 @@ TEST_CASE("Test Builtin Bytecode") {
 }
 
 
-TEST_CASE("Test List Comprehension Bytecode") {
+TEST_CASE(
+        "Test List Comprehension Bytecode"
+        ) {
     const std::string source = "list1 = [1, 2, 3]\nlistnew = [x for x in list1 if x < 3]";
     const std::vector expectedOpcodes = {PythonOpcode::RESUME, PythonOpcode::BUILD_LIST, PythonOpcode::LOAD_CONST,
                                          PythonOpcode::LIST_EXTEND, PythonOpcode::STORE_NAME, PythonOpcode::LOAD_NAME,
@@ -140,7 +150,9 @@ TEST_CASE("Test List Comprehension Bytecode") {
 }
 
 
-TEST_CASE("Test Simple Arithmetic Bytecode") {
+TEST_CASE(
+        "Test Simple Arithmetic Bytecode"
+        ) {
     const std::string source = "a = 1\nb = 2\nsummed = a + b\nprint(summed)";
     const std::vector expectedOpcodes = {PythonOpcode::RESUME, PythonOpcode::LOAD_SMALL_INT, PythonOpcode::STORE_NAME,
                                          PythonOpcode::LOAD_SMALL_INT, PythonOpcode::STORE_NAME,
@@ -154,12 +166,12 @@ TEST_CASE("Test Simple Arithmetic Bytecode") {
 
     REQUIRE(opcodes == expectedOpcodes);
 
-    const int* addendOne = std::get_if<int>(&(bytecodeModule.instructions[1].argval));
+    const int64_t* addendOne = std::get_if<int64_t>(&(bytecodeModule.instructions[1].argval));
     REQUIRE(addendOne != nullptr);
     REQUIRE(*addendOne == 1);
     REQUIRE(bytecodeModule.instructions[2].argrepr == "a");
 
-    const int* addendTwo = std::get_if<int>(&(bytecodeModule.instructions[3].argval));
+    const int64_t* addendTwo = std::get_if<int64_t>(&(bytecodeModule.instructions[3].argval));
     REQUIRE(addendTwo != nullptr);
     REQUIRE(*addendTwo == 2);
